refactor : 기존 레코드를 originalRecord로 명명
